some corrections

+++ projekat/demo/src/main/java/rs/team15/repository/TableRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import rs.team15.model.TableR;
 
 
-public interface TableRepository extends JpaRepository <TableR, Integer> {
+public interface TableRepository extends JpaRepository <TableR, Long> {
 }
